+++ pkg/signal/signal_test.go
@@ -13,7 +13,7 @@ import (
 
 func TestRandomSubset(t *testing.T) {
 	r := rand.New(testutil.RandSource(t))
-	base := FromRaw([]uint32{0, 1, 2, 3, 4}, 0)
+	base := FromRaw([]uint64{0, 1, 2, 3, 4}, 0)
 	var s Signal
 	for i := 0; i < 1000 && s.Len() < base.Len(); i++ {
 		delta := base.RandomSubset(r, 1)
@@ -24,18 +24,18 @@ func TestRandomSubset(t *testing.T) {
 }
 
 func TestSubtract(t *testing.T) {
-	base := FromRaw([]uint32{0, 1, 2, 3, 4}, 0)
+	base := FromRaw([]uint64{0, 1, 2, 3, 4}, 0)
 	assert.Equal(t, 5, base.Len())
-	base.Subtract(FromRaw([]uint32{0}, 0))
+	base.Subtract(FromRaw([]uint64{0}, 0))
 	assert.Equal(t, 4, base.Len())
-	base.Subtract(FromRaw([]uint32{1}, 0))
+	base.Subtract(FromRaw([]uint64{1}, 0))
 	assert.Equal(t, 3, base.Len())
 }
 
 func TestIntersectsWith(t *testing.T) {
-	base := FromRaw([]uint32{0, 1, 2, 3, 4}, 1)
-	assert.True(t, base.IntersectsWith(FromRaw([]uint32{0, 5, 10}, 1)))
-	assert.False(t, base.IntersectsWith(FromRaw([]uint32{5, 10, 15}, 1)))
+	base := FromRaw([]uint64{0, 1, 2, 3, 4}, 1)
+	assert.True(t, base.IntersectsWith(FromRaw([]uint64{0, 5, 10}, 1)))
+	assert.False(t, base.IntersectsWith(FromRaw([]uint64{5, 10, 15}, 1)))
 	// The other signal has a lower priority.
-	assert.False(t, base.IntersectsWith(FromRaw([]uint32{0, 1, 2}, 0)))
+	assert.False(t, base.IntersectsWith(FromRaw([]uint64{0, 1, 2}, 0)))
 }
